+++ doctr/__init__.py
@@ -1,5 +1,6 @@
 from .local import (encrypt_variable, encrypt_file, GitHub_post,
-    generate_GitHub_token, upload_GitHub_deploy_key, generate_ssh_key)
+    generate_GitHub_token, upload_GitHub_deploy_key, generate_ssh_key,
+    check_repo_exists)
 from .travis import (decrypt_file, setup_deploy_key, get_token, run,
     setup_GitHub_push, gh_pages_exists, create_gh_pages, sync_from_log,
     commit_docs, push_docs, get_current_repo, find_sphinx_build_dir)
@@ -7,6 +8,7 @@ from .travis import (decrypt_file, setup_deploy_key, get_token, run,
 __all__ = [
     'encrypt_variable', 'encrypt_file', 'GitHub_post',
     'generate_GitHub_token', 'upload_GitHub_deploy_key', 'generate_ssh_key',
+    'check_repo_exists',
 
     'decrypt_file', 'setup_deploy_key', 'get_token', 'run', 'setup_GitHub_push', 'gh_pages_exists',
     'create_gh_pages', 'sync_from_log', 'commit_docs', 'push_docs', 'get_current_repo', 'find_sphinx_build_dir'